Create UTLC.lean
With the help of PLFA, I wrote the framework of UTLC in LEAN4.

The UTLC in PLFA contains the following parts.
0. Define a Value-Term from existing Word(String).

1. Lambda Calculus
1.1 Value-Term (Exactly the part 0)

1.2 Lambda Lambda.
The application, which remain constant when the lhs can't acts on rhs. Otherwise, it can be reduced by substitution.

2. Natural Number with If-Else structure.
I leant that the If-Else structure based on Natural Number.
(P.S. 以前就听说 if else 是特殊的宏定义，从 PLFA 里学到了通过 Natural Number 定义的方法)

3. mu
Like the define syntax, the mu operator will try to evaluate the rhs term then define it as the lhs String.

And what need to be continued is Reduction to the inductive class Value.